# docs/delivery/out/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "АИскра — тренажёр оператора 112 / ДДС",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/v1/training/scenarios/generate": {
   "post": {
    "tags": [
     "training"
    ],
    "summary": "Сгенерировать сценарии (черновики на проверку)",
    "operationId": "generate_api_v1_training_scenarios_generate_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/GenerateIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/GeneratedOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/training/scenarios": {
   "get": {
    "tags": [
     "training"
    ],
    "summary": "Банк сценариев",
    "operationId": "list_scenarios_api_v1_training_scenarios_get",
    "parameters": [
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "name": "difficulty",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Difficulty"
      }
     },
     {
      "name": "card_type",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Card Type"
      }
     },
     {
      "name": "source",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Source"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 30,
       "title": "Page Size"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ScenarioPage"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/training/scenarios/{scenario_id}": {
   "get": {
    "tags": [
     "training"
    ],
    "summary": "Сценарий: легенда и эталоны",
    "operationId": "get_scenario_api_v1_training_scenarios__scenario_id__get",
    "parameters": [
     {
      "name": "scenario_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Scenario Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ScenarioView"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "training"
    ],
    "summary": "Правка сценария (возвращает в черновики)",
    "operationId": "edit_scenario_api_v1_training_scenarios__scenario_id__patch",
    "parameters": [
     {
      "name": "scenario_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Scenario Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/EditScenarioIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/StatusOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/training/scenarios/{scenario_id}/approve": {
   "post": {
    "tags": [
     "training"
    ],
    "summary": "Утвердить сценарий",
    "operationId": "approve_api_v1_training_scenarios__scenario_id__approve_post",
    "parameters": [
     {
      "name": "scenario_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Scenario Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/StatusOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/training/scenarios/{scenario_id}/archive": {
   "post": {
    "tags": [
     "training"
    ],
    "summary": "Сценарий в архив",
    "operationId": "archive_api_v1_training_scenarios__scenario_id__archive_post",
    "parameters": [
     {
      "name": "scenario_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Scenario Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/StatusOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/training/calls/incoming": {
   "post": {
    "tags": [
     "training"
    ],
    "summary": "Учебный входящий вызов в 112 (п. 1.4)",
    "operationId": "incoming_api_v1_training_calls_incoming_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/IncomingIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CallStarted"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/training/calls/{call_id}/answer": {
   "post": {
    "tags": [
     "training"
    ],
    "summary": "«Принять» — первая реплика заявителя",
    "operationId": "answer_api_v1_training_calls__call_id__answer_post",
    "parameters": [
     {
      "name": "call_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Call Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AnswerIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "anyOf": [
          {
           "$ref": "#/components/schemas/ReplicaOut"
          },
          {
           "type": "null"
          }
         ],
         "title": "Response Answer Api V1 Training Calls  Call Id  Answer Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/training/calls/dds": {
   "post": {
    "tags": [
     "training"
    ],
    "summary": "Звонок из АРМ ДДС: старшему группы, заявителю, в службу (п. 2.3)",
    "operationId": "dds_call_api_v1_training_calls_dds_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/DdsCallIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CallStarted"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/training/calls/{call_id}/replicas": {
   "post": {
    "tags": [
     "training"
    ],
    "summary": "Реплика оператора → ответ ИИ-собеседника",
    "operationId": "replica_api_v1_training_calls__call_id__replicas_post",
    "parameters": [
     {
      "name": "call_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Call Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ReplicaIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReplicaOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/training/speech": {
   "get": {
    "tags": [
     "training"
    ],
    "summary": "Доступен ли голосовой ввод (модель речи настроена)",
    "operationId": "speech_status_api_v1_training_speech_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SpeechStatusOut"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "training"
    ],
    "summary": "Распознать реплику в трубку (голосовой ввод, Whisper)",
    "operationId": "transcribe_api_v1_training_speech_post",
    "requestBody": {
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_transcribe_api_v1_training_speech_post"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SpeechOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/training/calls/{call_id}/end": {
   "post": {
    "tags": [
     "training"
    ],
    "summary": "Завершить звонок",
    "operationId": "end_call_api_v1_training_calls__call_id__end_post",
    "parameters": [
     {
      "name": "call_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Call Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/StatusOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "AnswerIn": {
    "properties": {
     "card_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Card Id"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "title": "AnswerIn"
   },
...
   "GenerateIn": {
    "properties": {
     "count": {
      "type": "integer",
      "maximum": 50.0,
      "minimum": 1.0,
      "title": "Count",
      "default": 5
     },
     "groups": {
      "items": {
       "type": "integer"
      },
      "type": "array",
      "title": "Groups",
      "description": "Группы классификатора; пусто — любые"
     },
     "difficulty": {
      "type": "integer",
      "maximum": 5.0,
      "minimum": 1.0,
      "title": "Difficulty",
      "default": 2
     }
    },
    "additionalProperties": false,
    "type": "object",
    "title": "GenerateIn"
   },
   "GeneratedOut": {
    "properties": {
     "ids": {
      "items": {
       "type": "string",
       "format": "uuid"
      },
      "type": "array",
      "title": "Ids"
     }
    },
    "type": "object",
    "required": [
     "ids"
    ],
    "title": "GeneratedOut"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ScenarioPage": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/ScenarioRow"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     }
    },
    "type": "object",
    "required": [
     "items",
     "total"
    ],
    "title": "ScenarioPage"
   },
...
   "StatusOut": {
    "properties": {
     "status": {
      "type": "string",
      "title": "Status"
     }
    },
    "type": "object",
    "required": [
     "status"
    ],
    "title": "StatusOut"
   },
...
  }
 }
}